En CuentaCorriente >> getSaldo() cambie lo que retornaba porque decia double y deberia retornar un float.

+++ trunk/src/sistema/CuentaCorriente.java
@@ -21,7 +21,7 @@ public class CuentaCorriente {
 		return movimientos;
 	}
 
-	public double getSaldo() {
+	public float getSaldo() {
 		return this.saldo;
 	}
